requirements.txt
asgiref==3.8.1
Django==4.2.8
django-seed==0.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Faker==33.1.0
iniconfig==2.0.0
packaging==24.2
pillow==10.2.0
pluggy==1.5.0
psycopg2==2.9.10
PyJWT==2.8.0
pytest==8.3.4
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sqlparse==0.5.3
toposort==1.10
typing_extensions==4.12.2
tzdata==2024.2
